Avoid error when deleting list.
The issue was that the `list.todos()` property access was getting evaluated in a separate computation (inside the each) from the check that the list existed. What this meant was that when the list was deleted there was a time where it was re-evaluated when the list did not exist. A moment later the not found got rendered, but it wasn't quick enough.

This could be related to #64 and perhaps solving that could mean we could revert this